Move PrettyPrinted to models module
